requirements.txt
# --- Flask & CORS ---
flask==2.3.3
flask-cors==4.0.0

# --- ChromaDB & Embeddings ---
chromadb==0.5.0
sentence-transformers==2.2.2

# --- Request Handling & Env ---
requests==2.31.0
python-dotenv==1.0.1

# --- OpenAI TTS ---
openai==1.14.3

# --- Audio Handling (optional, if streaming/processing) ---
pydub==0.25.1
